revert commit 227 folder structure changes
